Use gorm in problem history services

+++ model/problem_history.go
@@ -30,7 +30,7 @@ func (o Operation) String() string {
 // 题目历史记录
 type ProblemHistory struct {
 	Id           uint64            `gorm:"primaryKey;autoIncrement;comment:记录ID" json:"id,omitempty"`
-	UserId       int               `gorm:"not null;default:0;comment:用户ID" json:"user_id,omitempty"`
+	UserId       uint64            `gorm:"not null;default:0;comment:用户ID" json:"user_id,omitempty"`
 	ProblemId    uint64            `gorm:"not null;default:0;comment:题目ID" json:"problem_id,omitempty"`
 	Title        string            `gorm:"type:text;not null;comment:标题" json:"title,omitempty"`
 	Source       string            `gorm:"type:text;not null;comment:题目来源" json:"source,omitempty"`
